backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
pyjwt==2.8.0
dj-rest-auth==4.0.1
django-allauth==0.57.0
channels==4.0.0
channels-redis==4.1.0
celery==5.3.6
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
daphne==4.0.0
Pillow==10.1.0
django-filter==23.5